Add models/utils/db layers and refactor API into router modules (closes #10)
- src/api/db.py: shared DynamoDB table references (table, locations_table,
  subscribers_table) and constants (MAX_LIMIT, DEFAULT_LIMIT); tests patch
  db.table instead of app.table
- src/api/models.py: Lead, Location, Subscriber dataclasses with
  from_dynamo() and to_dict() (camelCase) methods
- src/api/utils.py: pure stateless helpers — parse_date, encode_key,
  decode_key, normalize_timestamp, now_iso
- src/api/routers/: leads, locations, subscribers, stripe Router modules;
  each uses Router from aws_lambda_powertools.event_handler.api_gateway
- src/api/app.py: slim entrypoint — creates APIGatewayRestResolver, includes
  all routers, exposes backward-compat shims (_transform_*, _parse_date etc.)
- scripts/local_api_server.py: replace importlib.spec_from_file_location with
  plain import app after inserting src/api/ onto sys.path
- src/scraper/dynamo.py: persist pdf_url field to DynamoDB (was extracted but
  never written)
- scripts/seed_local.py: carry pdf_url through CSV→DynamoDB seed path
- tests/fixtures/leads.py: add pdf_url to mock lead records
- tests/test_api.py, test_locations.py, test_subscribers.py: patch db.table /
  db.locations_table / db.subscribers_table instead of app.*; add import db
All 106 tests pass.
